[feature/#1011] create auth navigation
