+++ app/src/main/java/org/sopt/official/feature/auth/feature/authmain/navigation/AuthMainNavigation.kt
@@ -1,0 +1,23 @@
+package org.sopt.official.feature.auth.feature.authmain.navigation
+
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import kotlinx.serialization.Serializable
+import org.sopt.official.feature.auth.feature.authmain.AuthMainRoute
+
+fun NavGraphBuilder.authMainNavGraph(
+    navigateToUnAuthenticatedHome: () -> Unit,
+    onGoogleLoginCLick: () -> Unit,
+    navigateToCertification: () -> Unit
+) {
+    composable<AuthMainNavigation> {
+        AuthMainRoute(
+            navigateToUnAuthenticatedHome = navigateToUnAuthenticatedHome,
+            onGoogleLoginCLick = onGoogleLoginCLick,
+            navigateToCertification = navigateToCertification
+        )
+    }
+}
+
+@Serializable
+data object AuthMainNavigation
